switched floats to doubles. compiles.

+++ orbitSimulator/orbitSimulator/acceleration.cpp
@@ -4,7 +4,7 @@
  * SET ACCELERATION
  * Sets the speed and direction.
  *********************************************/
-void Acceleration::setAcc(float aGravity, const Direction &direction)
+void Acceleration::setAcc(double aGravity, const Direction &direction)
 {
 	setDDx(aGravity * direction.getDx());
 	setDDy(aGravity * direction.getDy());
