+++ orbitSimulator/orbitSimulator/testDirection.h
@@ -25,7 +25,7 @@ class TestDirection
 {
 public:
 	void run();
-	bool closeEnough(float actual, float expected, float tolerance);
+	bool closeEnough(double actual, double expected, double tolerance);
 	
 	void test_defaultConstructor();
 	void test_setUp();
